Move single_threaded to utilities.py
