+++ galsim/utilities.py
@@ -1862,3 +1862,33 @@ def get_omp_threads():
     :returns: num_threads
     """
     return _galsim.GetOMPThreads()
+
+class single_threaded:
+    """A context manager that turns off (or down) OpenMP threading e.g. during multiprocessing.
+
+    Parameters:
+        num_threads:    The number of threads you want during the context [default: 1]
+    """
+    def __init__(self, num_threads=1):
+        # Get the current number of threads here, so we can set it back when we're done.
+        self.orig_num_threads = get_omp_threads()
+        self.temp_num_threads = num_threads
+
+        # If threadpoolctl is installed, use that too, since it will set blas libraries to
+        # be single threaded too. This makes it so you don't need to set the environment
+        # variables OPENBLAS_NUM_THREAD=1 or MKL_NUM_THREADS=1, etc.
+        try:
+            import threadpoolctl
+        except ImportError:
+            self.tpl = None
+        else:  # pragma: no cover  (Not installed on GHA currently.)
+            self.tpl = threadpoolctl.threadpool_limits(num_threads)
+
+    def __enter__(self):
+        set_omp_threads(self.temp_num_threads)
+        return self
+
+    def __exit__(self, type, value, traceback):
+        set_omp_threads(self.orig_num_threads)
+        if self.tpl is not None:  # pragma: no cover
+            self.tpl.unregister()
